% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEITRNet.R
\name{compare_experiment_sets}
\alias{compare_experiment_sets}
\title{Compare Experiment Sets}
\usage{
compare_experiment_sets(experiment_sets_list)
}
\arguments{
\item{experiment_sets_list}{A list of experiment sets to compare}
}
\value{
A comparison of the experiment sets
}
\description{
This function compares the results of multiple experiment sets.
}
