% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEITRNet.R
\name{SEITR}
\alias{SEITR}
\title{SEITR Model}
\usage{
SEITR(t, state, parameters)
}
\arguments{
\item{t}{Time}

\item{state}{Initial state}

\item{parameters}{Model parameters}
}
\value{
A list of derivatives
}
\description{
This function defines the SEITR model for network analysis.
}
