% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEITRNet.R
\name{SEITR_network}
\alias{SEITR_network}
\title{SEITR Network Analysis}
\usage{
SEITR_network(
  network_type = "ER",
  n = 100,
  n_par1 = 0.9,
  n_par2 = 10,
  Lambda = 1.1,
  beta1 = 0.8,
  beta2 = 0.18,
  beta3 = 0.02,
  alpha1 = 0.1,
  alpha2 = 0.055,
  delta_I = 0.03,
  delta_T = 0.03,
  mu = 0.01,
  S = 85,
  E = 5,
  I = 10,
  Tt = 0,
  R = 0,
  N = n,
  t = 100,
  num_exp = 10,
  verbose = F
)
}
\arguments{
\item{network_type}{Type of network ("ER", "BA", "WS")}

\item{n}{Number of nodes}

\item{n_par1}{Parameter 1 for network generation}

\item{n_par2}{Parameter 2 for network generation}

\item{Lambda}{Birth rate}

\item{beta1}{Transmission rate}

\item{beta2}{Progression rate}

\item{beta3}{Recovery rate}

\item{alpha1}{Treatment rate}

\item{alpha2}{Recovery rate from treatment}

\item{delta_I}{Death rate from infection}

\item{delta_T}{Death rate from treatment}

\item{mu}{Natural death rate}

\item{S}{Initial susceptible population}

\item{E}{Initial exposed population}

\item{I}{Initial infected population}

\item{Tt}{Initial treatment population}

\item{R}{Initial recovered population}

\item{N}{Initial total population}

\item{t}{Time period}

\item{num_exp}{Number of experiments}

\item{verbose}{Verbose output}
}
\value{
A list of status counts
}
\description{
This function performs SEITR network analysis.
}
